requirements.txt
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost
# Use compatible MongoDB client (latest stable >=4.7)
pymongo>=4.7
dnspython
certifi
from_root
evidently==0.2.8
dill
PyYAML
neuro_mf
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
python-multipart

python-dotenv
